Simplify repository structure
We abandon the .zit.file/.git idea to switch to a structure where the Git dir
is just .file.git, and the working tree is the current directory.

This has the advantage of not requiring non-portable and fragile hardlinks,
although it requires a rather drastic .file.git/info/exclude file to simulate
the one-file repository.