Allow storage of repos inside .zit
Since tracking many files independently in the same directory would create a
large number of .*.git directories, we allow the use of a .zit subdirectory to
collect all repositories in.

The command 'zit init' without any arguments serves the purpose of creating
.zit/, and 'zit track somefile' will create .zit/somefile.git instead of
.somefile.git if .zit/ exists.

For all other commands, we do look for an existing .somefile.git if .zit/
exists but .zit/somefile.git does not.