Use current Zit or Git dir for exclude file
We put the generic exclude file in $ZIT_DIR, if it's present, or in $GIT_DIR
otherwise. This solution is more robust than the previous one relying on
~/.zitignore, since it doesn't set the config up with absolute paths, and it's
also cleaner as it doesn't create files outside of the Zit/Git dirs.

The diskspace saving is not as strong, since we still need at least one exclude
file per directory with tracked content, but it's still at least one block
cheaper than using the standard per-repo exclude file .git/info/exclude, since
we go for .somefile.git/exclude in the worst case.