zit: add new "zit add" command
"zit add" is a new command that either creates a new zit repository if
the file is unknown to zit, or simply does "git add" on the file if it
has already been tracked.

This makes it really easy for git users to use zit since they don't need
to learn about "zit track" anymore.